review fix: vpx definition
