+++ rust/cadeau/src/xmf/vpx/mod.rs
@@ -1,10 +1,10 @@
 use core::fmt;
 
 use xmf_sys::{
-    vpx::{XmfVpxCodecType, XmfVpxPacketKind},
-    XmfVpxFrame, XmfVpxFrame_Destroy, XmfVpxFrame_GetBuffer, XmfVpxFrame_GetDuration, XmfVpxFrame_GetFlags,
+    vpx::{XmfVpxCodecType, XmfVpxFrame, XmfVpxImage, XmfVpxPacket, XmfVpxPacketKind},
+    XmfVpxFrame_Destroy, XmfVpxFrame_GetBuffer, XmfVpxFrame_GetDuration, XmfVpxFrame_GetFlags,
     XmfVpxFrame_GetHeight, XmfVpxFrame_GetPartitionId, XmfVpxFrame_GetPts, XmfVpxFrame_GetSize,
-    XmfVpxFrame_GetSpatialLayerEncoded, XmfVpxFrame_GetWidth, XmfVpxImage, XmfVpxImage_Destroy, XmfVpxPacket,
+    XmfVpxFrame_GetSpatialLayerEncoded, XmfVpxFrame_GetWidth, XmfVpxImage_Destroy,
     XmfVpxPacket_Destroy, XmfVpxPacket_GetFrame, XmfVpxPacket_GetKind, XmfVpxPacket_IsEmpty,
 };
