+++ rust/xmf-sys/src/lib.rs
@@ -15,12 +15,7 @@ pub type XmfRecorder = c_void;
 
 pub type XmfWebMMuxer = c_void;
 
-pub type XmfVpxEncoder = c_void;
-pub type XmfVpxDecoder = c_void;
-pub type XmfVpxImage = c_void;
-pub type XmfVpxPacket = c_void;
-pub type XmfVpxFrame = c_void;
-pub type VpxIterator = *const c_void;
+
 
 pub const XMF_MUXER_FILE_OPEN_ERROR: c_int = -1001;
 pub const XMF_MUXER_PARSER_ERROR: c_int = -1002;
